IGNITE-24846 Move GridFunc to lang
