+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/F.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.GridFunc;
+import org.apache.ignite.internal.util.lang.GridFunc;
 
 /**
  * Defines {@code alias} for {@link GridFunc} by extending it. Since Java doesn't provide type aliases
