+++ modules/core/src/main/java/org/apache/ignite/internal/util/typedef/CX2.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.GridFunc;
+import org.apache.ignite.internal.util.lang.GridFunc;
 import org.apache.ignite.internal.util.lang.IgniteClosure2X;
 
 /**
