+++ modules/core/src/main/java/org/apache/ignite/internal/util/typedef/PX2.java
@@ -17,7 +17,7 @@
 
 package org.apache.ignite.internal.util.typedef;
 
-import org.apache.ignite.internal.util.GridFunc;
+import org.apache.ignite.internal.util.lang.GridFunc;
 import org.apache.ignite.internal.util.lang.IgnitePredicate2X;
 
 /**
